convert cargo tests to mocha
