+++ test/test-async.js
@@ -2148,275 +2148,6 @@ exports['whilst optional callback'] = function (test) {
 };
 
 
-
-exports['cargo'] = {
-
-    'cargo': function (test) {
-    test.expect(19);
-    var call_order = [],
-        delays = [160, 160, 80];
-
-    // worker: --12--34--5-
-    // order of completion: 1,2,3,4,5
-
-    var c = async.cargo(function (tasks, callback) {
-        setTimeout(function () {
-            call_order.push('process ' + tasks.join(' '));
-            callback('error', 'arg');
-        }, delays.shift());
-    }, 2);
-
-    c.push(1, function (err, arg) {
-        test.equal(err, 'error');
-        test.equal(arg, 'arg');
-        test.equal(c.length(), 3);
-        call_order.push('callback ' + 1);
-    });
-    c.push(2, function (err, arg) {
-        test.equal(err, 'error');
-        test.equal(arg, 'arg');
-        test.equal(c.length(), 3);
-        call_order.push('callback ' + 2);
-    });
-
-    test.equal(c.length(), 2);
-
-    // async push
-    setTimeout(function () {
-        c.push(3, function (err, arg) {
-            test.equal(err, 'error');
-            test.equal(arg, 'arg');
-            test.equal(c.length(), 1);
-            call_order.push('callback ' + 3);
-        });
-    }, 60);
-    setTimeout(function () {
-        c.push(4, function (err, arg) {
-            test.equal(err, 'error');
-            test.equal(arg, 'arg');
-            test.equal(c.length(), 1);
-            call_order.push('callback ' + 4);
-        });
-        test.equal(c.length(), 2);
-        c.push(5, function (err, arg) {
-            test.equal(err, 'error');
-            test.equal(arg, 'arg');
-            test.equal(c.length(), 0);
-            call_order.push('callback ' + 5);
-        });
-    }, 120);
-
-
-    setTimeout(function () {
-        test.same(call_order, [
-            'process 1 2', 'callback 1', 'callback 2',
-            'process 3 4', 'callback 3', 'callback 4',
-            'process 5'  , 'callback 5'
-        ]);
-        test.equal(c.length(), 0);
-        test.done();
-    }, 800);
-},
-
-    'without callback': function (test) {
-    test.expect(1);
-    var call_order = [],
-        delays = [160,80,240,80];
-
-    // worker: --1-2---34-5-
-    // order of completion: 1,2,3,4,5
-
-    var c = async.cargo(function (tasks, callback) {
-        setTimeout(function () {
-            call_order.push('process ' + tasks.join(' '));
-            callback('error', 'arg');
-        }, delays.shift());
-    }, 2);
-
-    c.push(1);
-
-    setTimeout(function () {
-        c.push(2);
-    }, 120);
-    setTimeout(function () {
-        c.push(3);
-        c.push(4);
-        c.push(5);
-    }, 180);
-
-    setTimeout(function () {
-        test.same(call_order, [
-            'process 1',
-            'process 2',
-            'process 3 4',
-            'process 5'
-        ]);
-        test.done();
-    }, 800);
-},
-
-    'bulk task': function (test) {
-    test.expect(7);
-    var call_order = [],
-        delays = [120,40];
-
-    // worker: -123-4-
-    // order of completion: 1,2,3,4
-
-    var c = async.cargo(function (tasks, callback) {
-        setTimeout(function () {
-            call_order.push('process ' + tasks.join(' '));
-            callback('error', tasks.join(' '));
-        }, delays.shift());
-    }, 3);
-
-    c.push( [1,2,3,4], function (err, arg) {
-        test.equal(err, 'error');
-        call_order.push('callback ' + arg);
-    });
-
-    test.equal(c.length(), 4);
-
-    setTimeout(function () {
-        test.same(call_order, [
-            'process 1 2 3', 'callback 1 2 3',
-            'callback 1 2 3', 'callback 1 2 3',
-            'process 4', 'callback 4',
-        ]);
-        test.equal(c.length(), 0);
-        test.done();
-    }, 800);
-},
-
-    'drain once': function (test) {
-    test.expect(1);
-
-    var c = async.cargo(function (tasks, callback) {
-        callback();
-    }, 3);
-
-    var drainCounter = 0;
-    c.drain = function () {
-        drainCounter++;
-    };
-
-    for(var i = 0; i < 10; i++){
-        c.push(i);
-    }
-
-    setTimeout(function(){
-        test.equal(drainCounter, 1);
-        test.done();
-    }, 500);
-},
-
-    'drain twice': function (test) {
-    test.expect(1);
-
-    var c = async.cargo(function (tasks, callback) {
-        callback();
-    }, 3);
-
-    var loadCargo = function(){
-        for(var i = 0; i < 10; i++){
-            c.push(i);
-        }
-    };
-
-    var drainCounter = 0;
-    c.drain = function () {
-        drainCounter++;
-    };
-
-    loadCargo();
-    setTimeout(loadCargo, 500);
-
-    setTimeout(function(){
-        test.equal(drainCounter, 2);
-        test.done();
-    }, 1000);
-},
-
-    'events': function(test) {
-    test.expect(4);
-    var calls = [];
-    var q = async.cargo(function(task, cb) {
-        // nop
-        calls.push('process ' + task);
-        async.setImmediate(cb);
-    }, 1);
-    q.concurrency = 3;
-
-    q.saturated = function() {
-        test.ok(q.length() == 3, 'cargo should be saturated now');
-        calls.push('saturated');
-    };
-    q.empty = function() {
-        test.ok(q.length() === 0, 'cargo should be empty now');
-        calls.push('empty');
-    };
-    q.drain = function() {
-        test.ok(
-            q.length() === 0 && q.running() === 0,
-            'cargo should be empty now and no more workers should be running'
-        );
-        calls.push('drain');
-        test.same(calls, [
-            'saturated',
-            'process foo',
-            'process bar',
-            'process zoo',
-            'foo cb',
-            'process poo',
-            'bar cb',
-            'empty',
-            'process moo',
-            'zoo cb',
-            'poo cb',
-            'moo cb',
-            'drain'
-        ]);
-        test.done();
-    };
-    q.push('foo', function () {calls.push('foo cb');});
-    q.push('bar', function () {calls.push('bar cb');});
-    q.push('zoo', function () {calls.push('zoo cb');});
-    q.push('poo', function () {calls.push('poo cb');});
-    q.push('moo', function () {calls.push('moo cb');});
-},
-
-    'expose payload': function (test) {
-    test.expect(5);
-    var called_once = false;
-    var cargo= async.cargo(function(tasks, cb) {
-        if (!called_once) {
-            test.equal(cargo.payload, 1);
-            test.ok(tasks.length === 1, 'should start with payload = 1');
-        } else {
-            test.equal(cargo.payload, 2);
-            test.ok(tasks.length === 2, 'next call shold have payload = 2');
-        }
-        called_once = true;
-        setTimeout(cb, 25);
-    }, 1);
-
-    cargo.drain = function () {
-        test.done();
-    };
-
-    test.equals(cargo.payload, 1);
-
-    cargo.push([1, 2, 3]);
-
-    setTimeout(function () {
-        cargo.payload = 2;
-    }, 15);
-}
-
-};
-
-
-
 exports['memoize'] = {
 
     'memoize': function (test) {
